Added the HttpCoreException class for custom error handling in HTTP connections. Added unit tests for use cases of the HttpCoreException class, for both NuNit and xUnit.
Renamed the classes for the different unit tests to match the names of the files that contain them. It is not necessary to have the word NuNit or xUnit in the class names, as each class is in a unit test project that already determines which unit test system will be used.