Added all the MarkDown files necessary to complete the first version, referring to how to work with the project. Includes:
- Code of Conduct.
- Changelog.
- Upcoming features file.
- How to make contributions.
- GPL v3 license file.
- Project security file.
- File showing how to use the class library.

Created a console project that allows generating pages for the project's Wiki.

Made changes to the preBuilding validation project so that it works correctly when we try to compile the class library project, verifying what was previously indicated.

The EncryptionsTests and FileManagementTest unit tests were renamed to EncryptionTests and FileManagementTests so that they can pass preBuilding validation.

Modified the YAML file responsible for performing preBuilding validation in GitHub Actions.